+++ R/imports.R
@@ -16,5 +16,6 @@
 #' @importFrom cli rule
 #' @importFrom workflowsets as_workflow_set
 #' @importFrom hardhat add_intercept_column
+#' @importFrom modeldata sim_classification
 
 NULL